Move all view-rendering controller specs to main project
