+++ publify_core/spec/controllers/authors_controller_spec.rb
@@ -12,8 +12,6 @@ describe AuthorsController, type: :controller do
       let!(:article) { create(:article, user: no_profile_user, published_at: now - 1.hour) }
 
       describe 'html' do
-        render_views
-
         before { get 'show', params: { id: no_profile_user.login } }
 
         it { expect(response).to render_template(:show) }
@@ -47,8 +45,6 @@ describe AuthorsController, type: :controller do
       let!(:article) { create(:article, user: full_profile_user) }
 
       describe 'html' do
-        render_views
-
         before { get 'show', params: { id: full_profile_user.login } }
 
         it { expect(response).to render_template(:show) }
