+++ publify_core/spec/controllers/notes_controller_spec.rb
@@ -3,7 +3,6 @@
 require 'rails_helper'
 
 describe NotesController, type: :controller do
-  render_views
   let!(:blog) { create(:blog) }
 
   describe '#index' do
